# src/api-gateway/controller/book.go
package controller

import (
	"api-gateway/dto"
	"api-gateway/helpers"
	bookpb "api-gateway/pb/bookpb"
	"fmt"
	"net/http"

	"github.com/golang/protobuf/ptypes/empty"
	"github.com/labstack/echo/v4"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
)

type BookController struct {
	Client bookpb.BookServiceClient
}

func NewBookController(client bookpb.BookServiceClient) BookController {
	return BookController{
		Client: client,
	}
}

// @Summary     Create a new book
// @Description Register a new book with the role 'Book'
// @Tags        books
// @Accept      json
// @Produce     json
// @Param       request body dto.BookRequest true "Book details"
// @Success     201 {object} dto.WebResponse
// @Failure     400 {object} dto.ErrResponse
// @Failure     500 {object} dto.ErrResponse
// @Router      /books [post]
func (b BookController) Create(c echo.Context) error {
	var bookRequest dto.BookRequest

	// Bind request body
	if err := c.Bind(&bookRequest); err != nil {
		return c.JSON(http.StatusBadRequest, map[string]string{"message": "invalid request parameters"})
	}

	// Validate request
	if err := validate.Struct(bookRequest); err != nil {
		return c.JSON(http.StatusBadRequest, map[string]string{"message": fmt.Sprintf("invalid request parameters: %v", err)})
	}

	ctx, cancel, err := helpers.NewServiceContext()
	if err != nil {
		return err
	}
	defer cancel()

	// Map to gRPC request
	grpcRequest := &bookpb.BookRequest{
		Title:         bookRequest.Title,
		Author:        bookRequest.Author,
		PublishedDate: bookRequest.PublishedDate,
		Status:        bookRequest.Status,
	}

	// Call gRPC service
	responseGrpc, err := b.Client.CreateBook(ctx, grpcRequest)
	if err != nil {
		if e, ok := status.FromError(err); ok {
			switch e.Code() {
			case codes.Internal:
				return c.JSON(http.StatusInternalServerError, map[string]string{"message": "Internal server error"})
			}
		}
		return c.JSON(http.StatusInternalServerError, map[string]string{"message": fmt.Sprintf("Error response: %v", err)})
	}

	data := helpers.UnmarshalJSONToWebResponse(responseGrpc.Data)

	return c.JSON(data.Status, data)
}

// @Summary     Get a book by ID
// @Description Retrieve a book by its unique ID
// @Tags        books
// @Accept      json
// @Produce     json
// @Param       id path string true "Book ID"
// @Success     200 {object} dto.BookResponse
// @Failure     404 {object} dto.ErrResponse
// @Failure     500 {object} dto.ErrResponse
// @Router      /books/{id} [get]
func (b BookController) GetBookById(c echo.Context) error {
	bookId := c.Param("id")

	ctx, cancel, err := helpers.NewServiceContext()
	if err != nil {
		return err
	}
	defer cancel()

	// Call gRPC service
	responseGrpc, err := b.Client.GetBookById(ctx, &bookpb.BookId{Id: bookId})
	if err != nil {
		if e, ok := status.FromError(err); ok {
			switch e.Code() {
			case codes.Internal:
				return c.JSON(http.StatusInternalServerError, map[string]string{"message": "Internal server error"})
			}
		}
		return c.JSON(http.StatusInternalServerError, map[string]string{"message": fmt.Sprintf("Error response: %v", err)})
	}

	data := helpers.UnmarshalJSONToWebResponse(responseGrpc.Data)

	return c.JSON(data.Status, data)
}

// @Summary     List all books
// @Description Retrieve all books
// @Tags        books
// @Accept      json
// @Produce     json
// @Success     200 {array} dto.BookResponse
// @Failure     500 {object} dto.ErrResponse
// @Router      /books [get]
func (b BookController) GetAllBooks(c echo.Context) error {
	ctx, cancel, err := helpers.NewServiceContext()
	if err != nil {
		return err
	}
	defer cancel()

	// Call gRPC service
	responseGrpc, err := b.Client.GetAllBooks(ctx, &empty.Empty{})
	if err != nil {
		if e, ok := status.FromError(err); ok {
			switch e.Code() {
			case codes.Internal:
				return c.JSON(http.StatusInternalServerError, map[string]string{"message": "Internal server error"})
			}
		}
		return c.JSON(http.StatusInternalServerError, map[string]string{"message": fmt.Sprintf("Error response: %v", err)})
	}

	data := helpers.UnmarshalJSONToWebResponse(responseGrpc.Data)

	return c.JSON(data.Status, data)
}

// @Summary     Update a book
// @Description Update a book's information by ID
// @Tags        books
// @Accept      json
// @Produce     json
// @Param       id path string true "Book ID"
// @Param       request body dto.BookRequest true "Updated book details"
// @Success     200 {object} dto.BookResponse
// @Failure     400 {object} dto.ErrResponse
// @Failure     500 {object} dto.ErrResponse
// @Router      /books/{id} [put]
func (b BookController) UpdateBookById(c echo.Context) error {
	bookId := c.Param("id")
	var bookRequest dto.BookRequest

	// Bind request body
	if err := c.Bind(&bookRequest); err != nil {
		return c.JSON(http.StatusBadRequest, map[string]string{"message": "invalid request parameters"})
	}

	// Validate request
	if err := validate.Struct(bookRequest); err != nil {
		return c.JSON(http.StatusBadRequest, map[string]string{"message": fmt.Sprintf("invalid request parameters: %v", err)})
	}

	ctx, cancel, err := helpers.NewServiceContext()
	if err != nil {
		return err
	}
	defer cancel()

	// Map to gRPC request
	grpcRequest := &bookpb.UpdateBookRequest{
		Title:         bookRequest.Title,
		Author:        bookRequest.Author,
		PublishedDate: bookRequest.PublishedDate,
		Status:        bookRequest.Status,
		BookId:        bookId,
	}

	// Call gRPC service
	responseGrpc, err := b.Client.UpdateBookById(ctx, grpcRequest)
	if err != nil {
		if e, ok := status.FromError(err); ok {
			switch e.Code() {
			case codes.Internal:
				return c.JSON(http.StatusInternalServerError, map[string]string{"message": "Internal server error"})
			}
		}
		return c.JSON(http.StatusInternalServerError, map[string]string{"message": fmt.Sprintf("Error response: %v", err)})
	}

	data := helpers.UnmarshalJSONToWebResponse(responseGrpc.Data)

	return c.JSON(data.Status, data)
}

// @Summary     Delete a book
// @Description Delete a book by its unique ID
// @Tags        books
// @Accept      json
// @Produce     json
// @Param       id path string true "Book ID"
// @Success     200 {object} dto.SuccessResponse
// @Failure     500 {object} dto.ErrResponse
// @Router      /books/{id} [delete]
func (b BookController) DeleteBookById(c echo.Context) error {
	bookId := c.Param("id")

	ctx, cancel, err := helpers.NewServiceContext()
	if err != nil {
		return err
	}
	defer cancel()

	// Call gRPC service
	responseGrpc, err := b.Client.DeleteBookById(ctx, &bookpb.BookId{Id: bookId})
	if err != nil {
		if e, ok := status.FromError(err); ok {
			switch e.Code() {
			case codes.Internal:
				return c.JSON(http.StatusInternalServerError, map[string]string{"message": "Internal server error"})
			}
		}
		return c.JSON(http.StatusInternalServerError, map[string]string{"message": fmt.Sprintf("Error response: %v", err)})
	}

	data := helpers.UnmarshalJSONToWebResponse(responseGrpc.Data)

	return c.JSON(data.Status, data)
}